[BUGFIX] Don't call DataObjectFactory::destroy() with nullptr.
- There were several places where nullptr was passed to DataObjectFactory::destroy().
- This doesn't lead to errors in the current code base, but:
  - it indicates that something does not work as expected by the programmer
  - it will lead to errors with a feature currently under development (data object reference counting)

+++ src/runtime/local/vectorized/Tasks.cpp
@@ -37,10 +37,11 @@ void CompiledPipelineTask<DenseMatrix<VT>>::execute(uint32_t fid, uint32_t batch
         accumulateOutputs(localResults, localAddRes, r, r2);
         
         // cleanup
-        for (auto &localResult : localResults) {
-            DataObjectFactory::destroy(localResult);
-            localResult = nullptr;
-        }
+        for (auto &localResult : localResults)
+            if(localResult) {
+                DataObjectFactory::destroy(localResult);
+                localResult = nullptr;
+            }
         this->cleanupFuncInputs(std::move(linputs));
     }
